requirements/requirements-prod.txt
PyYAML>=5.1,<6.0
lxml>=4.3.0,<5.0.0
musicbrainzngs==0.6
mutagen==1.42.0
urllib3>=1.24.0,<2.0.0
youtube-dl>=2019.3.18
requests>=2.21.0,<3.0.0
tqdm>=4.31.0,<5.0.0
bookmarks-parser>=0.1.3,<0.2.0
fire>=0.1.3,<0.2.0